backend/pyproject.toml
[tool.poetry]
name = "cowriter-backend"
version = "0.1.0"
description = "Backend for CoWriter application"
authors = ["CoWriter Team <github@cowriter.ai>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.8"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
pydantic = "^2.10.6"
openai = "^1.63.2"
requests = "^2.32.3"
sqlalchemy = "^2.0.29"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
email-validator = "^2.1.1"
asyncpg = "^0.29.0"
pydantic-settings = "^2.8.1"
aiosqlite = "^0.21.0"
greenlet = "^3.1.1"
pytest-asyncio = "^0.25.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
black = "^24.2.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
isort = "^5.13.2"
flake8-docstrings = "^1.7.0"
flake8-bugbear = "^24.2.6"
types-requests = "^2.32.0.20241016"
pytest-asyncio = "^0.25.3"
autoflake = "^2.3.1"
types-python-jose = "^3.4.0.20250224"
types-passlib = "^1.7.7.20241221"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
disable_error_code = ["valid-type", "misc"]

[[tool.mypy.overrides]]
module = "app.models.*"
disallow_subclassing_any = false

[[tool.mypy.overrides]]
module = "sqlalchemy.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "sqlalchemy.orm.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "sqlalchemy.ext.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "sqlalchemy.dialects.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
